+++ AddOns/LiteButtonAuras/CHANGELOG.md
@@ -1,6 +1,8 @@
 # LiteButtonAuras
 
-## [11.2.7-1](https://github.com/xod-wow/LiteButtonAuras/tree/11.2.7-1) (2025-12-03)
-[Full Changelog](https://github.com/xod-wow/LiteButtonAuras/compare/11.2.5-2...11.2.7-1) [Previous Releases](https://github.com/xod-wow/LiteButtonAuras/releases)
+## [11.2.7-2](https://github.com/xod-wow/LiteButtonAuras/tree/11.2.7-2) (2025-12-16)
+[Full Changelog](https://github.com/xod-wow/LiteButtonAuras/compare/11.2.7-1...11.2.7-2) [Previous Releases](https://github.com/xod-wow/LiteButtonAuras/releases)
 
-- Update ToC  
+- Fixes to blizzard mouseover/focuscast support  
+- Proper dynamic support for focuscast/mouseovercast #44  
+    Completely untested drycode, written from my sofa. Good luck to us all